Replace location by context in checkpoint indexer

+++ org.eclipse.linuxtools.tmf.core/src/org/eclipse/linuxtools/tmf/core/ctfadaptor/CtfLocation.java
@@ -64,7 +64,7 @@ public class CtfLocation implements ITmfLocation<Long>, Cloneable {
      */
     @Override
     public CtfLocation clone() {
-        return new CtfLocation(getLocation());
+        return new CtfLocation(getLocation().longValue());
     }
 
     /* (non-Javadoc)
